+++ src/clients/mailer.client.ts
@@ -1,10 +1,10 @@
 import { injectable, inject } from 'tsyringe';
 import nodemailer, { Transporter } from 'nodemailer';
 import type SMTPTransport from 'nodemailer/lib/smtp-transport';
-import { ILogger } from '../interfaces/logger.service.interface';
-import { TOKENS } from '../config/di.tokens';
+import { ILogger } from '../interfaces/logger-service.interface';
+import { TOKENS } from '../config/di-tokens.config';
 import ENV from '../config/env';
-import { IEmailTransport } from '../interfaces/email.client.interface';
+import { IEmailTransport } from '../interfaces/email-client.interface';
 
 @injectable()
 export class NodemailerTransport implements IEmailTransport {
