+++ src/container.ts
@@ -1,28 +1,28 @@
 import 'reflect-metadata';
 import { container } from 'tsyringe';
-import { TOKENS } from './config/di.tokens';
+import { TOKENS } from './config/di-tokens.config';
 import { RedisCacheService } from './services/cache.service';
-import { ICacheService } from './interfaces/cache.service.interface';
+import { ICacheService } from './interfaces/cache-service.interface';
 import { redisClient } from './clients/redis.client';
 import { NodemailerTransport } from './clients/mailer.client';
-import { IEmailTransport } from './interfaces/email.client.interface';
+import { IEmailTransport } from './interfaces/email-client.interface';
 import { GmailService } from './services/mail.service';
-import { IEmailService } from './interfaces/email.service.interface';
+import { IEmailService } from './interfaces/email-service.interface';
 import { SubscriptionService } from './services/subscription.service';
 import { SubscriptionController } from './controllers/subscription.controller';
 import { SubscriptionRepository } from './repositories/subscription.repository';
-import { ISubscriptionRepository } from './interfaces/subscription.repository.interface';
+import { ISubscriptionRepository } from './interfaces/subscription-repository.interface';
 import { WeatherService } from './services/weather.service';
 import { WeatherMapper } from './mappers/weather-data.mapper';
 import { WeatherDto } from './dto/weather.dto';
 import { WeatherController } from './controllers/weather.controller';
-import { WeatherDigestJob } from './jobs/weatherDigest.job';
-import { IWeatherMapper } from './interfaces/weather-data.mapper.interface';
-import { ILogger } from './interfaces/logger.service.interface';
+import { WeatherDigestJob } from './jobs/weather-digest.job';
+import { IWeatherMapper } from './interfaces/weather-data-mapper.interface';
+import { ILogger } from './interfaces/logger-service.interface';
 import { LoggerService } from './services/logger.service';
-import { ErrorHandlerMiddleware } from './middlewares/errorHandler';
-import { WeatherApiClient } from './clients/weatherApi.client';
-import { OpenWeatherMapClient } from './clients/openWeatherMap.client';
+import { ErrorHandlerMiddleware } from './middlewares/error-handler.middleware';
+import { WeatherApiClient } from './clients/weather-api.client';
+import { OpenWeatherMapClient } from './clients/open-weather-map.client';
 import ENV from './config/env';
 
 container.registerSingleton<ILogger>(TOKENS.ILogger, LoggerService);
