+++ src/interfaces/cache-service.interface.ts
