+++ src/interfaces/email-client.interface.ts
