+++ src/interfaces/email-send-options.interface.ts
