+++ src/interfaces/email-service.interface.ts
@@ -1,4 +1,4 @@
-import { IEmailSendOptions } from './email-send-options.interfaces';
+import { IEmailSendOptions } from './email-send-options.interface';
 
 export interface IEmailService {
   send(opts: IEmailSendOptions): Promise<void>;
