+++ src/interfaces/logger-service.interface.ts
