+++ src/interfaces/open-weather-map-error-response.interface.ts
