+++ src/interfaces/open-weather-map-response.interface.ts
