+++ src/interfaces/subscription-repository.interface.ts
