+++ src/interfaces/weather-api-error-response.interface.ts
