+++ src/interfaces/weather-data-mapper.interface.ts
