+++ src/interfaces/weather-provider.interface.ts
