+++ src/jobs/scheduler.job.ts
@@ -1,7 +1,7 @@
 import cron from 'node-cron';
 import { container } from '../container';
-import { WeatherDigestJob } from './weatherDigest.job';
-import CRON_FREQUENCY from '../utils/constants/cron.constants';
+import { WeatherDigestJob } from './weather-digest.job';
+import CRON_FREQUENCY from '../utils/constants/cron.constant';
 
 const job = container.resolve(WeatherDigestJob);
