+++ src/tests/setup/global-teardown.setup.ts
@@ -2,5 +2,5 @@ import { execSync } from 'child_process';
 
 module.exports = async () => {
   const root = process.cwd();
-  execSync('npx ts-node src/scripts/drop-test-db.ts', { cwd: root, stdio: 'inherit' });
+  execSync('npx ts-node src/scripts/drop-test-db.script.ts', { cwd: root, stdio: 'inherit' });
 };
