+++ src/tests/unit/weather-mapper.test.ts
