+++ src/utils/constants/weather-api-error-code.constant.ts
