+++ src/utils/gen-token.util.ts
